requirements.txt
beautifulsoup4==4.5.1
python-dateutil==2.6.0
six==1.10.0
wheel==0.24.0
